feat: create es2015 modules and build app with webpack
