+++ public/index.html
@@ -33,7 +33,7 @@
   </section>
 </footer>
 
-<script src="./bundle.js" type="module"></script>
+<script src="./bundle.js"></script>
 
 </body>
 </html>
