+++ src/components/templates/show-more.js
@@ -1,0 +1,3 @@
+export const createTemplateBtnShowMore = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
