+++ src/components/templates/sort.js
@@ -1,0 +1,13 @@
+export const createTemplateSort = () => {
+  return `
+    <ul class="sort">
+      <li>
+        <a href="#" class="sort__button sort__button--active">
+          Sort by default
+        </a>
+      </li>
+      <li><a href="#" class="sort__button">Sort by date</a></li>
+      <li><a href="#" class="sort__button">Sort by rating</a></li>
+    </ul>
+  `;
+};
