+++ src/components/templates/statistics.js
@@ -1,0 +1,3 @@
+export const createTemplateStatistics = () => {
+  return `<p>130 291 movies inside</p>`;
+};
